Remboursement mensuel de l'Assurance Maladie par departement

Extrait de 100 observations issues du fichier original de la base R202001 comportant 4 298 838 lignes

Usage

R202001

Format

A data frame with 100 rows and 39 variables:

PRS_FAC_TOP

Pour la prise en charge des nouveaux travailleurs independants dans le SNDS à partir du 1er janvier 2019, l’identification des factures se fera par la reutilisation des codes petits regimes suivants : RGM_COD = 105, 106, 107, 103, 360, 660, chacun de ces petits regimes permettant de distinguer les differentes categories d’independants. De plus, les codes des organismes d’affiliation et de liquidation ne feront plus reference au code grand regime = 03, mais prendront le code grand regime = 01. Un top permettant d’identifier les factures de ces nouveaux travailleurs independants est cree.

l_fac_top

Libelle des factures des nouveaux travailleurs independants

act_coe

coefficient global

act_dnb

dénombrement

asu_nat

nature d'assurance

ben_qlt

qualité du bénéficiaire

cpam

cpam d'affiliation du bénéficiaire

cpl_cod

complement d'acte

dep_mon

dépassement d'honoraires

dpt

département d'affiliation du bénéficiaire

exe_spe

spécialité exécutant

exe_spe1

regroupement spécialité exécutant

exe_stj1

mode d'exercice exécutant

l_asu_nat

libellé assurance

l_ben_qlt

libellé qualité du bénéficiaire

l_cpam

libellé cpam d'affiliation du bénéficiaire

l_cpl_cod

libellé du complément d'acte

l_dpt

libellé du département d'affiliation du bénéficiaire

l_exe_spe

libellé spécialité exécutant

l_exe_spe1

libellé regroupement spécialité exécutant

l_exe_stj1

libellé mode d'exercice exécutant

l_pre_spe

libellé spécialité prescripteur

l_pre_spe1

libelle regroupement spécialité prescripteur

l_pre_stj1

libellé mode d'exercice prescripteur

l_prs_nat

libellé prestation

l_region

libellé région d'affiliation du bénéficiaire

l_serie

libellé de la série Statistique Mensuelle

l_top_slm

libellé Top Sections Locales Mutualistes

pre_spe

spécialité prescripteur

pre_spe1

regroupement spécialité prescripteur

pre_stj1

mode d'exercice prescripteur

prs_nat

prestation

rec_mon

base de remboursement

region

région d'affiliation du bénéficiaire

rem_mon

montant remboursé

rem_tau

taux de remboursement

serie

série de la Statistique Mensuelle

sns_date

mois et année des soins

top_slm

Top Sections Locales Mutualistes

Source

data.gouv.fr

Références

SNIIRAM (2020) (AMELI)