Creation d'une base de nombre d'actes par praticien et par departement

Creation d'une base de nombre d'actes par praticien et par departement

Usage

mt_crea_nb_actes_par_praticien(
  chemin,
  nom_fichier_dep,
  nom_fichier_reg,
  ecriture = T
)

Arguments

chemin

Une chaine de caractere: Chemin des donnees brutes des deux fichiers xlsx d'AMELI

nom_fichier_dep

Une chaine de caractere correspondant au nom du fichier departemental xlsx

nom_fichier_reg

Une chaine de caractere correspondant au nom du fichier regional xlsx

ecriture

Booleen Par defaut TRUE. Permet d'ecrire dans le dossier initial. Le desactiver n'a d'interet que pour les tests

Valeur de retour

NULL. Cette fonction cree lune base de nombre d'actes par praticien et par departement a l'emplacement initial

Exemples

chemin = system.file("extdata",package = "medtRucks")
nom_fichier_dep = "2020_patientele-des-medecins-liberaux-ape-par-departement_serie-annuelle.xlsx"
nom_fichier_reg = "2020_patientele-des-medecins-liberaux-ape-par-region_serie-annuelle.xlsx"
mt_crea_nb_actes_par_praticien(chemin = chemin,
                      nom_fichier_dep = nom_fichier_dep,
                      nom_fichier_reg=nom_fichier_reg,
                      ecriture = FALSE)